requirements.txt
pandas>=1.5.0
pymysql>=1.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
